+++ src/main/java/LottoApplication.java
@@ -1,6 +1,6 @@
 import controller.LottoController;
 
-public class Application {
+public class LottoApplication {
 
     public static void main(String[] args) {
         try {
