+++ src/main/java/model/Lotto.java
@@ -1,4 +1,4 @@
-package domain;
+package model;
 
 import java.util.*;
