+++ src/main/java/model/LottoNumber.java
@@ -1,4 +1,4 @@
-package domain;
+package model;
 
 import java.util.Objects;
 import java.util.Random;
