+++ src/main/java/model/LottoNumbers.java
@@ -1,4 +1,4 @@
-package domain;
+package model;
 
 import java.util.*;
